refactor: extract native app drag bridge hook from MainLayout
Made-with: Cursor
